# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/roles": {
            "get": {
                "description": "获取角色列表（支持分页、搜索、排序）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色"
                ],
                "summary": "获取角色列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码，默认1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量，默认10，最大100",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "搜索关键词（搜索角色名称）",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序字段",
                        "name": "order_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序方式：asc/desc",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-github_com_lyj404_gin-api-template_domain_dto_PaginationResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "创建新角色",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色"
                ],
                "summary": "创建角色",
                "parameters": [
                    {
                        "description": "角色信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_dto.CreateRoleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-github_com_lyj404_gin-api-template_domain_dto_RoleResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-string"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-string"
                        }
                    }
                }
            }
        },
        "/roles/:id": {
            "get": {
                "description": "根据ID获取角色详情",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色"
                ],
                "summary": "获取角色详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "角色ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-github_com_lyj404_gin-api-template_domain_dto_RoleResponse"
                        }
                    },
                    "404": {
                        "description": "角色不存在",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-string"
                        }
                    }
                }
            },
            "put": {
                "description": "更新角色信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色"
                ],
                "summary": "更新角色",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "角色ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "角色信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_dto.UpdateRoleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-github_com_lyj404_gin-api-template_domain_dto_RoleResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-string"
                        }
                    },
                    "404": {
                        "description": "角色不存在",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-string"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-string"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除角色",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色"
                ],
                "summary": "删除角色",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "角色ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-string"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-string"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_lyj404_gin-api-template_domain_dto.CreateRoleRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "github_com_lyj404_gin-api-template_domain_dto.PaginationResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "数据列表"
                },
                "page": {
                    "description": "当前页码",
                    "type": "integer"
                },
                "page_size": {
                    "description": "每页数量",
                    "type": "integer"
                },
                "total": {
                    "description": "总记录数",
                    "type": "integer"
                },
                "total_page": {
                    "description": "总页数",
                    "type": "integer"
                }
            }
        },
...
        "github_com_lyj404_gin-api-template_domain_dto.RoleResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_system": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "github_com_lyj404_gin-api-template_domain_dto.UpdateRoleRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "github_com_lyj404_gin-api-template_domain_result.ResponseResult-github_com_lyj404_gin-api-template_domain_dto_PaginationResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "@Description 响应状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "@Description 响应数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_dto.PaginationResponse"
                        }
                    ]
                },
                "message": {
                    "description": "@Description 响应信息",
                    "type": "string"
                }
            }
        },
...
        "github_com_lyj404_gin-api-template_domain_result.ResponseResult-github_com_lyj404_gin-api-template_domain_dto_RoleResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "@Description 响应状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "@Description 响应数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_dto.RoleResponse"
                        }
                    ]
                },
                "message": {
                    "description": "@Description 响应信息",
                    "type": "string"
                }
            }
        },
...
        "github_com_lyj404_gin-api-template_domain_result.ResponseResult-string": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "@Description 响应状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "@Description 响应数据",
                    "type": "string"
                },
                "message": {
                    "description": "@Description 响应信息",
                    "type": "string"
                }
            }
        },
...
    }
}